+++ nanomesh/image2mesh/__init__.py
@@ -1,9 +1,11 @@
-from .mesher2d import Mesher2D, plane2mesh
-from .mesher3d import Mesher3D, volume2mesh
+from ._mesher import Mesher
+from ._mesher2d import Mesher2D, plane2mesh
+from ._mesher3d import Mesher3D, volume2mesh
 
 __all__ = [
     'Mesher2D',
     'Mesher3D',
+    'Mesher',
     'plane2mesh',
     'volume2mesh',
 ]
