+++ nanomesh/image2mesh/_mesher2d/_polygon.py
