+++ nanomesh/image2mesh/_mesher3d/_bounding_box.py
